uan: Fix precision loss in Eb/N0 calculation in UanPhyPerCommonModes::CalcPer() (reported by donghuiyong)

+++ src/uan/model/uan-phy-gen.cc
@@ -310,7 +310,7 @@ UanPhyPerCommonModes::CalcPer(Ptr<Packet> pkt, double sinrDb, UanTxMode mode)
     // Amplitude Modulation", EE 242 Digital Communications and Codings, 2009
     case UanTxMode::QAM: {
         // generic EbNo
-        EbNo *= mode.GetBandwidthHz() / mode.GetDataRateBps();
+        EbNo *= static_cast<double>(mode.GetBandwidthHz()) / mode.GetDataRateBps();
 
         auto M = (double)mode.GetConstellationSize();
